Separate some of the windows font code from IGraphicsWin (.h/,cpp) to allow reuse
